added multiple pages using react-router-dom
