+++ eatsafe/package.json
@@ -12,6 +12,7 @@
     "@types/react-dom": "^18.2.19",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "^5.0.1",
     "react-webcam": "^7.2.0",
     "tesseract.js": "^5.0.4",
