+++ eatsafe/src/AboutUs.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import Header from "./components/Header";
+import MissionStatement from "./components/AboutComponets/MissionStatement";
+
+function AboutUs() {
+  return (
+    <div className="App">
+      <Header />
+      <MissionStatement />
+    </div>
+  );
+}
+
+export default AboutUs;
