+++ eatsafe/src/components/AboutComponets/MissionStatement.tsx
@@ -1,0 +1,11 @@
+import React, { useState } from 'react';
+import './AboutUs.css';
+
+function MissionStatement() {
+    return (
+        <div>
+        </div>
+    );
+}
+
+export default MissionStatement;
